refactor(statistics): 调整状态码类型为u16并优化索引配置
- 将StatisticsRequestStatusCode中的status_code类型从i64改为u16
- 更新request_status_count任务中状态码处理逻辑，统一转换为u16类型
- 优化request_log序列化配置，跳过None值字段
- 新增aiway-logs和request-logs两个索引的yaml配置文件
- 配置request-logs索引字段映射及时间戳分区策略
- 设置索引提交超时时间和分片目标文档数

+++ console/src/server/db/models/statistics_request_status_code.rs
@@ -9,7 +9,7 @@ use rocket::serde::{Deserialize, Serialize};
 #[builder(default)]
 pub struct StatisticsRequestStatusCode {
     /// 状态码
-    pub status_code: Option<i64>,
+    pub status_code: Option<u16>,
     ///  数量
     pub count: Option<i64>,
     /// 分钟起始时间戳（秒，0秒），包含，范围为`[state_time, state_time+59]`
